# public/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Rest API",
        "contact": {
            "name": "Md Amjad hossain",
            "url": ""
        },
        "version": "1.0"
    },
    "host": "localhost:8080/",
    "basePath": "/",
    "schemes": [
        "http"
    ],
    "paths": {
        "api/household": {
            "get": {
                "tags": [
                    "HouseHoldActivities"
                ],
                "summary": "Find All HouseHoldActivitiess",
                "description": "Returns HouseHoldActivities overview.",
                "operationId": "index",
                "produces": [
                    "application/xml",
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "HouseHoldActivities overview."
                    },
                    "401": {
                        "description": "Unauthorized action."
                    }
                }
            },
            "post": {
                "tags": [
                    "HouseHoldActivities"
                ],
                "summary": "Save HouseHoldActivitiess",
                "description": "",
                "operationId": "saveHouseHoldActivities",
                "consumes": [
                    "application/json",
                    "application/xml"
                ],
                "produces": [
                    "application/xml",
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "description": "HouseHoldActivitiess that needs to be added to the store",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/HouseHoldActivities"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "HouseHoldActivities overview."
                    },
                    "401": {
                        "description": "Unauthorized action."
                    }
                }
            }
        },
        "api/household/{household_unique_code}": {
            "get": {
                "tags": [
                    "HouseHoldActivities"
                ],
                "summary": "Find All HouseHoldActivitiess",
                "description": "Returns HouseHoldActivities overview.",
                "operationId": "index",
                "produces": [
                    "application/xml",
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "household_unique_code",
                        "in": "path",
                        "description": "household_unique_code to filter by",
                        "required": true,
                        "type": "string",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "multi"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "HouseHoldActivities overview."
                    },
                    "401": {
                        "description": "Unauthorized action."
                    }
                }
            }
        },
        "api/household/{id}": {
            "delete": {
                "tags": [
                    "HouseHoldActivities"
                ],
                "summary": "Delete HouseHoldActivities",
                "description": "",
                "operationId": "deleteHouseHoldActivities",
                "consumes": [
                    "application/xml",
                    "application/json",
                    "multipart/form-data",
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/xml",
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "HouseHoldActivities id to delete",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Invalid pet value"
                    }
                }
            }
        },
        "api/household/characteristics/{household_unique_code}": {
            "put": {
                "tags": [
                    "HouseHoldActivities"
                ],
                "summary": "Update an existing HouseHoldActivities",
                "description": "",
                "operationId": "updateHouseHoldActivities",
                "consumes": [
                    "application/json",
                    "application/xml"
                ],
                "produces": [
                    "application/xml",
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "household_unique_code",
                        "in": "path",
                        "description": "household_unique_code to filter by",
                        "required": true,
                        "type": "string",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "multi"
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "description": "HouseHoldActivities object that needs to be added to the store[c01-c11]",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/HouseHoldActivities"
                        }
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Invalid household_unique_code supplied"
                    },
                    "404": {
                        "description": "HouseHold not found"
                    },
                    "405": {
                        "description": "Validation exception"
                    }
                },
                "security": [
                    {
                        "petstore_auth": [
                            "write:pets",
                            "read:pets"
                        ]
                    }
                ]
            }
        },
        "api/household/injurymorbidity/{household_unique_code}": {
            "put": {
                "tags": [
                    "HouseHoldActivities"
                ],
                "summary": "Update an existing HouseHoldActivities",
                "description": "",
                "operationId": "updateHouseHoldActivities",
                "consumes": [
                    "application/json",
                    "application/xml"
                ],
                "produces": [
                    "application/xml",
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "household_unique_code",
                        "in": "path",
                        "description": "household_unique_code to filter by",
                        "required": true,
                        "type": "string",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "multi"
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "description": "HouseHoldActivities object that needs to be added to the store[e01-e33]",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/HouseHoldActivities"
                        }
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Invalid household_unique_code supplied"
                    },
                    "404": {
                        "description": "HouseHold not found"
                    },
                    "405": {
                        "description": "Validation exception"
                    }
                },
                "security": [
                    {
                        "petstore_auth": [
                            "write:pets",
                            "read:pets"
                        ]
                    }
                ]
            }
        },
        "api/household/suicideattemptactivity/{household_unique_code}": {
            "put": {
                "tags": [
                    "HouseHoldActivities"
                ],
                "summary": "Update an existing HouseHoldActivities",
                "description": "",
                "operationId": "updateHouseHoldActivities",
                "consumes": [
                    "application/json",
                    "application/xml"
                ],
                "produces": [
                    "application/xml",
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "household_unique_code",
                        "in": "path",
                        "description": "household_unique_code to filter by",
                        "required": true,
                        "type": "string",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "multi"
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "description": "HouseHoldActivities object that needs to be added to the store[[g01-g03]]",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/HouseHoldActivities"
                        }
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Invalid household_unique_code supplied"
                    },
                    "404": {
                        "description": "HouseHold not found"
                    },
                    "405": {
                        "description": "Validation exception"
                    }
                },
                "security": [
                    {
                        "petstore_auth": [
                            "write:pets",
                            "read:pets"
                        ]
                    }
                ]
            }
        },
        "api/household/deathconfirmation/{household_unique_code}": {
            "put": {
                "tags": [
                    "HouseHoldActivities"
                ],
                "summary": "Update an existing HouseHoldActivities",
                "description": "",
                "operationId": "updateHouseHoldActivities",
                "consumes": [
                    "application/json",
                    "application/xml"
                ],
                "produces": [
                    "application/xml",
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "household_unique_code",
                        "in": "path",
                        "description": "household_unique_code to filter by",
                        "required": true,
                        "type": "string",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "multi"
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "description": "HouseHoldActivities object that needs to be added to the store[d01-d07]",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/HouseHoldActivities"
                        }
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Invalid household_unique_code supplied"
                    },
                    "404": {
                        "description": "HouseHold not found"
                    },
                    "405": {
                        "description": "Validation exception"
                    }
                },
                "security": [
                    {
                        "petstore_auth": [
                            "write:pets",
                            "read:pets"
                        ]
                    }
                ]
            }
        },
...
    },
...
}